Exit with status 1 if config check fails, and fix terminating when reading send sockets

+++ keepalived/core/pidfile.c
@@ -34,6 +34,7 @@
 #include "pidfile.h"
 #include "main.h"
 #include "bitops.h"
+#include "utils.h"
 
 const char *pid_directory = PID_DIR PACKAGE;
 
@@ -59,7 +60,12 @@ int
 pidfile_write(const char *pid_file, int pid)
 {
 	FILE *pidfile = NULL;
-	int pidfd = creat(pid_file, S_IRUSR | S_IWUSR | S_IRGRP | S_IROTH);
+	int pidfd;
+
+	if (__test_bit(CONFIG_TEST_BIT, &debug))
+		return 1;
+
+	pidfd = creat(pid_file, S_IRUSR | S_IWUSR | S_IRGRP | S_IROTH);
 	if (pidfd != -1) pidfile = fdopen(pidfd, "w");
 
 	if (!pidfile) {
@@ -76,6 +82,9 @@ pidfile_write(const char *pid_file, int pid)
 void
 pidfile_rm(const char *pid_file)
 {
+	if (__test_bit(CONFIG_TEST_BIT, &debug))
+		return;
+
 	unlink(pid_file);
 }
 
@@ -117,6 +126,9 @@ process_running(const char *pid_file)
 bool
 keepalived_running(unsigned long mode)
 {
+	if (__test_bit(CONFIG_TEST_BIT, &debug))
+		return false;
+
 	if (process_running(main_pidfile))
 		return true;
 #ifdef _WITH_VRRP_
